Tweaked Particle Manager and added functionality to create multiple instances of an Emittor. Cars now display effect when colliding and removed all warnings.

+++ Trunk/Battle Cars/Battle Cars/Source/CBullet.cpp
@@ -38,7 +38,7 @@ void CBullet::Render(CCamera* camera)
 	temppos.right = (LONG)(temppos.left + GetWidth());
 	temppos.bottom = (LONG)(temppos.top + GetHeight());
 	CSGD_Direct3D::GetInstance()->DrawRect(temppos,255,255,255);
-	m_pTM->Draw(GetImageID(),(int)GetPosX() - camera->GetCamX() + camera->GetRenderPosX(),(int)GetPosY() - camera->GetCamY() + camera->GetRenderPosY(),m_fScale,m_fScale);
+	m_pTM->Draw(GetImageID(),(int)GetPosX() - (int)camera->GetCamX() + (int)camera->GetRenderPosX(),(int)GetPosY() - (int)camera->GetCamY() + (int)camera->GetRenderPosY(),m_fScale,m_fScale);
 	//m_pTM->Draw(GetImageID(),(int)GetPosX(),(int)GetPosY(),0.5f,0.5f,NULL,GetPosX()-(GetWidth()),GetPosY()-(GetHeight()),m_fRotation);
 }
